pyproject.toml
[tool.poetry]
name = "sync-tool"
version = "0.0.1"
description = ""
authors = ["Pascal Helmer <pascal@pascalhelmer.de>"]
readme = "README.md"
packages = [{ include = "sync_tool" }]

[tool.poetry.scripts]
sync-tool = "sync_tool.cli:cli"

[tool.poetry.plugins."sync.tool.provider"]
sync-tool-provider-jama = "sync_tool.providers.jama:JamaProvider"
sync-tool-provider-azure-devops = "sync_tool.providers.azure_dev_ops:AzureDevOpsProvider"
sync-tool-provider-testing = "sync_tool.providers.testing:TestingProvider"

[tool.poetry.extras]
performance = ["uvloop"]
jama = ["py-jama-rest-client"]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
pydantic = "2.7.1"
uvloop = { version = "0.19.0", optional = true }
structlog = "24.1.0"
py-jama-rest-client = { version = "1.17.1", optional = true }
typer-slim = "0.12.3"
python-dotenv = "1.0.1"
azure-devops = "^7.1.0b4"
types-requests = "^2.31.0.20240406"

[tool.poetry.group.dev.dependencies]
pre-commit = "3.7.0"
flake8 = "7.0.0"
black = "24.4.0"
isort = "5.13.2"
mypy = "1.9.0"
pytest = "8.1.1"
pytest-asyncio = "0.23.6"
pytest-cov = "5.0.0"
python-semantic-release = "9.4.2"
sphinx = "7.2.6"
sphinx-rtd-theme = "2.0.0"
sphinx-autodoc-typehints = "2.0.1"
autodoc-pydantic = "2.1.0"
livereload = "2.6.3"
pytest-watch = "4.2.0"
pytest-sugar = "1.0.0"
pytest-icdiff = "0.9"
bandit = { extras = ["toml"], version = "1.7.8" }
pytest-mock = "3.14.0"
pyfakefs = "5.4.1"
enum-tools = { extras = ["sphinx"], version = "0.12.0" }
py-jama-rest-client = "1.17.1"

[build-system]
requires = ["poetry-core>=1.6.1"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120
skip = "tests"

[tool.bandit]
exclude_dirs = ["docs", "tests"]
skips = []

[tool.semantic_release]
build_command = "source $VENV && poetry build && cd .."
commit_author = "github-actions <github-actions@github.com>"
commit_message = "chore(release): releasing version {version}  [skip ci]"
major_on_zero = true
version_variables = [
    "sync_tool/__init__.py:__version__"
]
version_toml = [
    "pyproject.toml:tool.poetry.version"
]

[tool.semantic_release.branches.stable]
match = "main"
prerelease = false

[tool.semantic_release.branches.beta]
match = "develop"
prerelease = true
prerelease_token = "rc"

[tool.semantic_release.branches.alpha]
match = "feature-.*"
prerelease = true
prerelease_token = "alpha"

[tool.semantic_release.changelog]
exclude_commit_patterns = [
    "^Merge pull request",
    "^Merge remote",
    "^Merge branch",
    "^chore(release):.*",
    "^build(deps-dev):.*"
]
